requirements.txt
openai
python-dotenv
requests
colorama
selenium
webdriver-manager
Pillow
undetected-chromedriver
moviepy==1.0.3
